♻️ Refaktorer rammevedtakstabell-kode for bedre lesbarhet
- Del opp lagRammevedtakstabellPrivatBil.ts i tre fokuserte filer under rammevedtak/
- Fjern unødvendig behandling-parameter fra lagRammevedtakstabell
- Introduser formaterKrVerdi i rammevedtakstabellUtils for å eliminere duplisert formateringslogikk
- Bytt const arrow functions til function-deklarasjoner for hjelpefunksjoner
- Bruk formaterIsoPeriodeMedTankestrek direkte i stedet for lokal formaterPeriode-wrapper
- Fiks navneinkonsistens: lagRammevedtaktabellUtenDelperioder → lagRammevedtakstabellUtenDelperioder
