+++ src/frontend/komponenter/Brev/Brevmeny.tsx
@@ -15,8 +15,8 @@ import { MellomlagretBrevDto, parseMellomlagretBrev } from './mellomlagring';
 import { lagInnvilgetPerioderPunktliste } from './punktliste/lagInnvilgetPerioderPunktliste';
 import { lagVerdier } from './stønadsverdier/lagVerdier';
 import { Fritekst, FritekstAvsnitt, MalStruktur, Tekst, Valg, Valgfelt } from './typer';
-import { lagRammevedtakstabell } from './vedtakstabell/lagRammevedtakstabell';
 import { lagVedtakstabell } from './vedtakstabell/lagVedtakstabell';
+import { lagRammevedtakstabell } from './vedtakstabell/rammevedtak/lagRammevedtakstabell';
 import { useApp } from '../../context/AppContext';
 import { usePersonopplysninger } from '../../context/PersonopplysningerContext';
 import { Behandling } from '../../typer/behandling/behandling';
@@ -144,7 +144,7 @@ export const Brevmeny: React.FC<Props> = ({
                 vedtak
             ),
             [variabelBeregningstabellId]: lagVedtakstabell(behandling, vedtak),
-            [variabelRammevedtaktabellReiseMedBil]: lagRammevedtakstabell(behandling, vedtak),
+            [variabelRammevedtaktabellReiseMedBil]: lagRammevedtakstabell(vedtak),
         };
         return htmlVariabler;
     }
