+++ src/frontend/komponenter/Brev/vedtakstabell/rammevedtak/rammevedtakstabellUtils.ts
@@ -1,0 +1,3 @@
+export function formaterKrVerdi(verdi: string | number, bekreftet: boolean): string {
+    return bekreftet ? `${verdi} kr` : `${verdi} kr*`;
+}
